+++ backend/controllers/ModeratorApprovalRequestsController.php
@@ -2,6 +2,7 @@
 
 namespace backend\controllers;
 
+use Yii;
 use backend\models\ModeratorApprovalRequests;
 use backend\models\ModeratorApprovalRequestsSearch;
 use yii\web\Controller;
@@ -38,9 +39,43 @@ class ModeratorApprovalRequestsController extends Controller
      */
     public function actionIndex()
     {
+        // Only a moderator or system admin can view a student request
+		if (Yii::$app->user->isGuest) {
+			Yii::$app->session->setFlash('error', 'Access allowed after login.');
+            return $this->goHome();
+        }
+		
+		if(!isset($_SESSION['user_id'])){
+				Yii::$app->session->setFlash('error', 'Session Expired. Please login again.');
+				Yii::$app->user->logout();
+				return $this->goHome();
+		}
+		
+		if(!isset($_SESSION['user_is'])){
+				Yii::$app->session->setFlash('error', 'Access allowed to moderator and system admins only.');
+				return $this->goHome();
+		}
+		
+		if($_SESSION['user_is'] != 'moderator'&&$_SESSION['user_is'] != 'systemadmin'){
+				Yii::$app->session->setFlash('error', 'Access allowed to moderator and system admins only.');
+				return $this->goHome();
+		}
+        //-------------------------------------------
+        $courseid = Yii::$app->request->get('course_id');
+        $userid = Yii::$app->session->get('user_id');
+        $useris = Yii::$app->session->get('user_is');
+        if($courseid!==null){
+            Yii::$app->session->set('course_code',$courseid);
+        }
         $searchModel = new ModeratorApprovalRequestsSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
 
+        if ($courseid !== null) {
+            if($useris=='moderator')
+            $dataProvider->query->andFilterWhere(['course_id' => $courseid,'moderator_id'=>$userid]);
+            else
+            $dataProvider->query->andFilterWhere(['course_id' => $courseid,'admin_id'=>$userid]);
+        }
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
@@ -67,6 +102,29 @@ class ModeratorApprovalRequestsController extends Controller
      */
     public function actionCreate()
     {
+        
+        // Only a moderator can create a moderator request
+		if (Yii::$app->user->isGuest) {
+			Yii::$app->session->setFlash('error', 'Access allowed after login.');
+            return $this->goHome();
+        }
+		
+		if(!isset($_SESSION['user_id'])){
+				Yii::$app->session->setFlash('error', 'Session Expired. Please login again.');
+				Yii::$app->user->logout();
+				return $this->goHome();
+		}
+		
+		if(!isset($_SESSION['user_is'])){
+				Yii::$app->session->setFlash('error', 'Access allowed to moderators only.');
+				return $this->goHome();
+		}
+		
+		if($_SESSION['user_is'] != 'moderator'){
+				Yii::$app->session->setFlash('error', 'Access allowed to moderators only.');
+				return $this->goHome();
+		}
+        //-------------------------------------------
         $model = new ModeratorApprovalRequests();
         $model->status='pending';
         $model->date_sent=date('Y-m-d H:i:s');
@@ -95,6 +153,28 @@ class ModeratorApprovalRequestsController extends Controller
      */
     public function actionUpdate($idModerator_Approval_Requests)
     {
+        // Only a system admin can update status of a moderator request
+		if (Yii::$app->user->isGuest) {
+			Yii::$app->session->setFlash('error', 'Access allowed after login.');
+            return $this->goHome();
+        }
+		
+		if(!isset($_SESSION['user_id'])){
+				Yii::$app->session->setFlash('error', 'Session Expired. Please login again.');
+				Yii::$app->user->logout();
+				return $this->goHome();
+		}
+		
+		if(!isset($_SESSION['user_is'])){
+				Yii::$app->session->setFlash('error', 'Access allowed to system admin only.');
+				return $this->goHome();
+		}
+		
+		if($_SESSION['user_is'] != 'systemadmin'){
+				Yii::$app->session->setFlash('error', 'Access allowed to system admin only.');
+				return $this->goHome();
+		}
+        //-------------------------------------------
         $model = $this->findModel($idModerator_Approval_Requests);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -115,6 +195,28 @@ class ModeratorApprovalRequestsController extends Controller
      */
     public function actionDelete($idModerator_Approval_Requests)
     {
+        // Only a system admin can delete a moderator request
+		if (Yii::$app->user->isGuest) {
+			Yii::$app->session->setFlash('error', 'Access allowed after login.');
+            return $this->goHome();
+        }
+		
+		if(!isset($_SESSION['user_id'])){
+				Yii::$app->session->setFlash('error', 'Session Expired. Please login again.');
+				Yii::$app->user->logout();
+				return $this->goHome();
+		}
+		
+		if(!isset($_SESSION['user_is'])){
+				Yii::$app->session->setFlash('error', 'Access allowed to system admin only.');
+				return $this->goHome();
+		}
+		
+		if($_SESSION['user_is'] != 'systemadmin'){
+				Yii::$app->session->setFlash('error', 'Access allowed to system admin only.');
+				return $this->goHome();
+		}
+        //-------------------------------------------
         $this->findModel($idModerator_Approval_Requests)->delete();
 
         return $this->redirect(['index']);
